+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 ## Version 2.1.16 (28 Jun 2021)
 
 * fix logging of deprecated args [manthey]
+* add shepards example [tourtiere]
 
 ## Version 2.1.15 (27 Dec 2020)